+++ xskillscore/core/comparative.py
@@ -77,29 +77,31 @@ def sign_test(
 
     Examples
     --------
-    >>> np.random.seed(42)
     >>> f1 = xr.DataArray(np.random.normal(size=(30)),
-    ...      coords=[('time', np.arange(30))])
-    >>> f2 = xr.DataArray(np.random.normal(size=(30)),
-    ...      coords=[('time', np.arange(30))]) + 2.
+    ...                   coords=[('time', np.arange(30))])
+    >>> f2 = f1 + 2
     >>> o = xr.DataArray(np.random.normal(size=(30)),
-    ...      coords=[('time', np.arange(30))])
-    >>> significantly_different, walk, confidence = sign_test(f1, f2, o,
-            time_dim='time', metric='mae', orientation='negative')
-    >>> walk.plot()
-    >>> confidence.plot(color='gray')
-    >>> (-1*confidence).plot(color='gray')
+    ...                  coords=[('time', np.arange(30))])
+    >>> significantly_different, walk, confidence = xs.sign_test(
+    ... f1, f2, o, time_dim='time', metric='mae', orientation='negative'
+    ... )
+    >>> walk.plot() # doctest: +ELLIPSIS
+    [<matplotlib.lines.Line2D object at 0x...>]
+    >>> confidence.plot(color='gray') # doctest: +ELLIPSIS
+    [<matplotlib.lines.Line2D object at 0x...>]
+    >>> (-1 * confidence).plot(color='gray') # doctest: +ELLIPSIS
+    [<matplotlib.lines.Line2D object at 0x...>]
     >>> walk
     <xarray.DataArray (time: 30)>
-    array([ 1,  2,  3,  2,  3,  2,  3,  4,  3,  4,  5,  4,  5,  4,  5,  4,  5,
-            6,  7,  8,  9, 10,  9, 10, 11, 12, 11, 12, 13, 14])
+    array([ 1,  0,  1,  2,  1,  2,  3,  4,  5,  6,  5,  6,  7,  6,  7,  8,  9,
+           10,  9, 10, 11, 12, 13, 12, 11, 12, 13, 14, 15, 14])
     Coordinates:
       * time     (time) int64 0 1 2 3 4 5 6 7 8 9 ... 20 21 22 23 24 25 26 27 28 29
     >>> significantly_different
     <xarray.DataArray (time: 30)>
     array([False, False, False, False, False, False, False, False, False,
-           False, False, False, False, False, False, False, False, False,
-           False, False,  True,  True, False,  True,  True,  True,  True,
+           False, False, False, False, False, False,  True,  True,  True,
+            True,  True,  True,  True,  True,  True,  True,  True,  True,
             True,  True,  True])
     Coordinates:
       * time     (time) int64 0 1 2 3 4 5 6 7 8 9 ... 20 21 22 23 24 25 26 27 28 29
@@ -243,15 +245,15 @@ def mae_test(
 
     Examples
     --------
-    >>> np.random.seed(42)
     >>> f1 = xr.DataArray(np.random.normal(size=(30)),
-    ...      coords=[('time', np.arange(30))])
+    ...                   coords=[('time', np.arange(30))])
     >>> f2 = xr.DataArray(np.random.normal(size=(30)),
-    ...      coords=[('time', np.arange(30))])
+    ...                   coords=[('time', np.arange(30))])
     >>> o = xr.DataArray(np.random.normal(size=(30)),
-    ...      coords=[('time', np.arange(30))])
-    >>> significantly_different, diff, hwci = mae_test(f1, f2, o, time_dim='time',
-            dim=[], alpha=0.05)
+    ...                  coords=[('time', np.arange(30))])
+    >>> significantly_different, diff, hwci = xs.mae_test(
+    ...    f1, f2, o, time_dim='time', dim=[], alpha=0.05
+    ... )
     >>> significantly_different
     <xarray.DataArray ()>
     array(False)
@@ -264,8 +266,9 @@ def mae_test(
     >>> # absolute magnitude of difference is smaller than half-width of
     >>> # confidence interval, therefore not significant at level alpha=0.05
     >>> # now comparing against an offset f2, the difference in MAE is significant
-    >>> significantly_different, diff, hwci = mae_test(f1, f2 + 2., o, time_dim='time',
-            dim=[], alpha=0.05)
+    >>> significantly_different, diff, hwci = xs.mae_test(
+    ... f1, f2 + 2., o, time_dim='time', dim=[], alpha=0.05
+    ... )
     >>> significantly_different
     <xarray.DataArray ()>
     array(True)
@@ -274,7 +277,6 @@ def mae_test(
     References
     ----------
         * https://www.cawcr.gov.au/projects/verification/CIdiff/FAQ-CIdiff.html
-
     """
     if isinstance(dim, str):
         dim = [dim]